Add a new Smart Contract to the SHARE smart contract protocol suite that
is an evolution of S2RD. In this new contract liquidity of royalty splits
will be enabled. This allows for the trading of revenue split percentages
through the inheritance of the ERC-721 standard. A test file is also
coupled to ensure basic reliability of the contract's functionality. The
Immutable contract includes a new data element and a new array of these
Uint256 elements. All other changes are related to deployment and
testing.